requirements.txt
openai>=1.30.0
tavily-python>=0.3.3
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.2
pywebview>=5.0
pyinstaller>=6.6.0
